create_flask_service/template/requirements.txt
flask==1.1.1
flask-restx==0.2.0
flask_accepts==0.16.3
flask-compress==1.4.0
flask-cors==3.0.8
python-dotenv==0.12.0
marshmallow==3.5.1
python-jose==3.1.0
werkzeug==1.0.0
flask-sqlalchemy==1.3.15
pyodbc==4.0.30
typing-extensions==3.7.4.1